Added TerraformCommonInputs class and modified testcases.
