+++ cs-hashicorp-terraform/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    * (c) Copyright 2019 Micro Focus
+    * (c) Copyright 2020 Micro Focus
     * All rights reserved. This program and the accompanying materials
     * are made available under the terms of the Apache License v2.0 which accompany this distribution.
     *
@@ -78,15 +78,15 @@
         <maven-deploy-plugin.version>2.8.2</maven-deploy-plugin.version>
         <maven-gpg-plugin.version>1.6</maven-gpg-plugin.version>
         <!--Dependencies versions-->
-        <cs-commons.version>0.0.7-RC1</cs-commons.version>
+        <cs-commons.version>0.0.6</cs-commons.version>
         <score-content-sdk.version>1.10.7</score-content-sdk.version>
         <junit.version>4.12</junit.version>
         <common-codec.version>1.10</common-codec.version>
         <cs-http-client.version>0.1.75</cs-http-client.version>
         <!--Misc properties-->
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.7</maven.compiler.source>
-        <maven.compiler.target>1.7</maven.compiler.target>
+        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>1.8</maven.compiler.target>
         <maven.sources.skip>true</maven.sources.skip>
         <custom.nexus>overwritten_by_ci</custom.nexus>
     </properties>
@@ -239,7 +239,7 @@
                         <include>pom.xml</include>
                     </includes>
                     <properties>
-                        <copyright.year>2019</copyright.year>
+                        <copyright.year>2020</copyright.year>
                     </properties>
                     <useDefaultMapping>false</useDefaultMapping>
                     <mapping>
