+++ cs-hashicorp-terraform/src/main/java/io/cloudslang/content/hashicorp/terraform/services/HttpCommons.java
@@ -12,23 +12,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-
 package io.cloudslang.content.hashicorp.terraform.services;
 
-import io.cloudslang.content.hashicorp.terraform.utils.Inputs;
+import io.cloudslang.content.hashicorp.terraform.entities.TerraformCommonInputs;
 import io.cloudslang.content.httpclient.entities.HttpClientInputs;
 import org.jetbrains.annotations.NotNull;
 
-import static io.cloudslang.content.hashicorp.terraform.utils.HttpUtils.setConnectionParameters;
-import static io.cloudslang.content.hashicorp.terraform.utils.HttpUtils.setProxy;
-import static io.cloudslang.content.hashicorp.terraform.utils.HttpUtils.setSecurityInputs;
+import static io.cloudslang.content.hashicorp.terraform.utils.HttpUtils.*;
 
 public class HttpCommons {
 
-    @NotNull
     static void setCommonHttpInputs(@NotNull final HttpClientInputs httpClientInputs,
-                                    @NotNull final Inputs commonInputs) {
+                                    @NotNull final TerraformCommonInputs commonInputs) {
         setProxy(httpClientInputs,
                 commonInputs.getProxyHost(),
                 commonInputs.getProxyPort(),
