+++ cs-hashicorp-terraform/src/main/java/io/cloudslang/content/hashicorp/terraform/utils/Constants.java
@@ -27,11 +27,8 @@ public class Constants {
         public static final String STRICT = "strict";
         public static final String EXCEPTION_NULL_EMPTY = "The %s can't be null or empty.";
         public static final String EXCEPTION_INVALID_PROXY = "The %s is not a valid proxy details.";
-        public static final String EXCEPTION_INVALID_LOGIN_TYPE = "The %s must be either 'API' or 'Native'.";
-        public static final String EXCEPTION_INVALID_LOGIN_TYPE_REST = "The %s or %s is required for login.";
         public static final String EXCEPTION_INVALID_BOOLEAN = "The %s for %s input is not a valid boolean value.";
         public static final String EXCEPTION_INVALID_NUMBER = "The %s for %s input is not a valid number value.";
-        public static final String EXCEPTION_EMPTY_FILE_PATH_AND_CONTENT_BYTES = "The filePath or both contentName and contentBytes inputs are required.";
         public static final String ANONYMOUS = "anonymous";
         public static final String GET = "GET";
         public static final String POST = "POST";
@@ -51,7 +48,6 @@ public class Constants {
         public static final String TERRAFORM_HOST = "app.terraform.io";
         public static final String TERRAFORM_VERSION_CONSTANT = "0.12.1";
         public static final String ORGANIZATION_PATH = "/organizations/";
-        public static final String ORGANIZATION_NAME = "organizationName";
         public static final String PATH_SEPARATOR = "/";
         public static final String AND = "&";
         public static final String QUERY = "?";
